refactor: move ConnectivityState inside models directory
